+++ components/auth/login-form.tsx
@@ -55,7 +55,7 @@ export const LoginForm = () => {
         <div className={cn('flex h-screen bg-gray-100')}>
             <div className={cn('flex flex-col justify-center items-center w-1/2 bg-white p-10')}>
                 <Image
-                    src='/logo.png'
+                    src='https://vnruzfzvnvdhb848.public.blob.vercel-storage.com/public/logo-Xm0liee8ZmpclDqhTaYAi5vKcHPiEz.png'
                     alt='logo'
                     width={200}
                     height={38}
@@ -151,7 +151,7 @@ export const LoginForm = () => {
             <div className={cn('relative w-full md:w-1/2 lg:w-1/2 mt-10 md:mt-0 flex justify-center md:justify-center flex-col items-center bg-gray-900 text-white')}>
                 <div className={cn('relative w-[30rem] h-[30rem]')}>
                     <div className={cn('absolute inset-0 rounded-full border-[16px] border-gray-800')}></div>
-                    <Image src="/trainer_1.png"
+                    <Image src="https://vnruzfzvnvdhb848.public.blob.vercel-storage.com/public/trainer_1-3XsmrTsq0RrcTtYZpXgf2EcVP4cyw6.png"
                         width={500}
                         height={500}
                         alt="Entrenador"